Move some methods to modules folder, add back button to game_init view
